+++ cypress.config.js
@@ -5,7 +5,6 @@ module.exports = defineConfig({
     setupNodeEvents (on, config) {
       // implement node event listeners here
     },
-    baseUrl: 'http://localhost:8080',
-    includeShadowDom: true
+    baseUrl: 'http://localhost:8000'
   }
 })