+++ cypress/e2e/cursor.cy.js
@@ -7,16 +7,16 @@ describe('Cursor in the player', () => {
     })
     cy.get('player-controls').contains('0.5 s')
     let startTime
-    cy.get('player-controls').find('#play-button').click().then(() => {
+    cy.get('#play-button').click().then(() => {
       startTime = Date.now()
     })
-      .find('.pause-icon').should('be.visible')
+    cy.get('player-controls').find('.pause-icon').should('be.visible')
     cy.get('player-controls').find('.play-icon').should('be.visible')
       .then(() => {
         const endTime = Date.now()
         const elapsed = endTime - startTime
         // FIXME: measuring the time this way is unreliable and often this fails
-        expect(elapsed).to.gt(400).and.to.lt(600)
+        expect(elapsed).to.gt(300).and.to.lt(500)
       })
   })
 
@@ -28,6 +28,7 @@ describe('Cursor in the player', () => {
     cy.get('audio-player').then(el => {
       el.get(0).setDuration(0.6)
     })
+    cy.get('player-controls').contains('0.6 s')
     cy.get('player-controls').contains('0.3 s')
     // Estimate how much time it takes for the cursor to travel across
     let startTime
@@ -40,7 +41,7 @@ describe('Cursor in the player', () => {
         const endTime = new Date()
         const elapsed = endTime - startTime
         // FIXME: measuring the time this way is unreliable and often this fails
-        expect(elapsed).to.gt(200).and.to.lt(400)
+        expect(elapsed).to.gt(150).and.to.lt(400)
       })
   })
 })