+++ src/elements/sound-element.ts
@@ -1,5 +1,10 @@
-import { createShadow, funDummy } from './utils.js'
-import dragScrollOrClick from '../events/drag-and-drop.js'
+import { parseTemplate, funDummy } from './utils'
+import dragScrollOrClick from '../events/drag-and-drop'
+
+import template from '../templates/sound-element.html'
+import '../style/sound-element.css'
+
+const nodes = parseTemplate(template)
 
 interface SoundParams {
   soundName: string
@@ -10,51 +15,45 @@ interface SoundParams {
 }
 
 export class SoundElement extends HTMLElement {
-  soundName: string
-  descriptionFile: string
+  soundName: string = ''
+  descriptionFile: string = ''
   audioFile?: string
   buttonCallback: (sound: SoundElement) => void = funDummy
   showDescription: (sound: SoundElement) => void = funDummy
   width?: number // Width on the player, set the first time the sound enters a track
-  connect: () => void
-  connected: Promise<void>
-  audio: HTMLAudioElement = document.createElement('audio')
-  addButton: HTMLButtonElement = document.createElement('button')
-
-  static fromParams (params: SoundParams): SoundElement {
-    let soundElement = new SoundElement()
-    soundElement = Object.assign(soundElement, params)
-    soundElement.initialize()
-    return soundElement
-  }
+  audio: HTMLAudioElement
+  addButton: HTMLButtonElement
 
   constructor () {
     super()
-    createShadow('sound-element', this, () => { this.connectEvents() })
-    this.soundName = ''
-    this.descriptionFile = ''
-    this.connect = () => { throw new Error('Impossible!') }
-    this.connected = new Promise(resolve => {
-      this.connect = resolve
-    })
+    // When we clone the sound, this will be run again even though the child exists
+    if (this.children.length === 0) {
+      this.appendChild(nodes.cloneNode(true))
+    }
+    this.addButton = this.querySelector('button') as HTMLButtonElement
+    this.audio = this.querySelector('audio') as HTMLAudioElement
     this.tabIndex = 0
-    this.setAttribute('tabindex', '0')
   }
 
-  initialize (): void {
+  initialize (params: SoundParams): void {
     this.classList.add('initializing')
-    this.append(this.soundName)
+
+    this.soundName = params.soundName
+    this.descriptionFile = params.descriptionFile
+    this.audioFile = params.audioFile
+    this.buttonCallback = params.buttonCallback
+    this.showDescription = params.showDescription
+
+    const span = document.createElement('span')
+    span.append(this.soundName)
+    this.append(span)
     this.id = `sound-${this.soundName}`.replace('ä', 'a').replace('ö', 'o')
-    this.connected
-      .then(() => {
-        this.connectAudio()
-        this.classList.remove('initializing')
-      })
-      .catch((error) => { throw error })
+    this.connectAudio()
+    this.connectEvents()
+    this.classList.remove('initializing')
   }
 
   connectEvents (): void {
-    this.addButton = this.shadowRoot?.querySelector('button') as HTMLButtonElement
     this.addButton.addEventListener('pointerdown', (event) => {
       event.stopPropagation()
       this.buttonCallback(this)
@@ -63,15 +62,14 @@ export class SoundElement extends HTMLElement {
     this.addEventListener('focus', () => { this.showDescription(this) })
     if (this.audioFile != null) {
       this.addEventListener('pointerdown', (event) => {
-        dragScrollOrClick(event, this)
+        if (event.pointerType !== 'mouse' || event.button === 0) {
+          dragScrollOrClick(event, this)
+        }
       }, { passive: true })
     }
-    this.connect()
   }
 
   connectAudio (): void {
-    const audio = this.shadowRoot?.querySelector('audio') as HTMLAudioElement
-    this.audio = audio
     if (this.audioFile != null) {
       this.audio.src = this.audioFile
       this.classList.add('with-audio')