+++ src/index.ts
@@ -1,3 +1,0 @@
-import './elements/app.js'
-
-document.title = 'Ääninen'